fixes to work with recams
they have leading nop messages, and the server info message isn't the
second message in the block.  In addition, the player entity isn't the
view entity in a recam, so allow manually specifying the player entity.

This could be made more robust but they're just some quick fixes to get
recams working.